refactor(extensions): split the tenancy tripwire out of loader.ts (#2466)
The hardening pushed loader.ts to 608 lines mixing two unrelated concerns: route
mounting + the default-deny auth guard, and verifying what an extension's
migrations did to the database. Split per the CLAUDE.md guideline (split when a
file mixes unrelated concerns, not merely on line count).

  loader.ts            608 -> 195   (mounting, auth guard)
  tenancyTripwire.ts        -> 415  (catalog reconciliation, RLS assertions)

Pure move, no behavior change: same 54 unit + 18 real-Postgres tests pass. Also
drops an orphaned duplicate docstring left over from the previous commit.

Refs #2466
